+++ packages/loaders/json/src/index.ts
@@ -1,3 +1,3 @@
-import {JsonFileLoader} from '@graphql-toolkit/json-file-loader';
+import {JsonFileLoader} from '@graphql-tools/json-file-loader';
 
 export default new JsonFileLoader();